Add support for disabling theme autoload
By default, if a theme exists in vim-clap that matches the name of the
current theme, it will be used no matter what. Setting a string of
"default", or any other unknown theme, causes the "material_design_dark"
theme to be used, *not* the default theme.

This change removes setting the `s:palette` variable to the
`material_design_dark#palette` when the `g:clap_theme` is invalid.
Instead it leaves the `s:palette` blank so that the default (adaptive)
theme is applied.